db/mongodb.py
"""MongoDB connector implementation (stub for future use)."""
from typing import Dict, Any, Optional
from .base import BaseConnector


class MongoDBConnector(BaseConnector):
    """MongoDB connector - to be implemented when needed."""
    
    def connect(self) -> None:
        """Establish connection to MongoDB."""
        raise NotImplementedError("MongoDB connector not yet implemented. Install pymongo and implement.")
    
    def disconnect(self) -> None:
        """Close MongoDB connection."""
        raise NotImplementedError("MongoDB connector not yet implemented")
    
    def fetch_last_row(self, table_name: str, order_by: Optional[str] = None) -> Dict[str, Any]:
        """Fetch the last document from a MongoDB collection."""
        raise NotImplementedError("MongoDB connector not yet implemented")
    
    def test_connection(self) -> bool:
        """Test if MongoDB connection is working."""
        raise NotImplementedError("MongoDB connector not yet implemented")


# Example implementation when you need it:
"""
from pymongo import MongoClient

class MongoDBConnector(BaseConnector):
    def connect(self) -> None:
        self.connection = MongoClient(
            host=self.config['host'],
            port=self.config.get('port', 27017),
            username=self.config.get('username'),
            password=self.config.get('password')
        )
        self.db = self.connection[self.config['database']]
    
    def fetch_last_row(self, table_name: str, order_by: Optional[str] = None) -> Dict[str, Any]:
        collection = self.db[table_name]
        sort_field = order_by if order_by else '_id'
        doc = collection.find_one(sort=[(sort_field, -1)])
        if doc and '_id' in doc:
            doc['_id'] = str(doc['_id'])  # Convert ObjectId to string
        return doc or {}
"""
