db/mysql.py
"""MySQL database connector implementation (stub for future use)."""
from typing import Dict, Any, Optional
from .base import BaseConnector


class MySQLConnector(BaseConnector):
    """MySQL connector - to be implemented when needed."""
    
    def connect(self) -> None:
        """Establish connection to MySQL database."""
        raise NotImplementedError("MySQL connector not yet implemented. Install pymysql/mysql-connector-python and implement.")
    
    def disconnect(self) -> None:
        """Close MySQL connection."""
        raise NotImplementedError("MySQL connector not yet implemented")
    
    def fetch_last_row(self, table_name: str, order_by: Optional[str] = None) -> Dict[str, Any]:
        """Fetch the last row from a MySQL table."""
        raise NotImplementedError("MySQL connector not yet implemented")
    
    def test_connection(self) -> bool:
        """Test if MySQL connection is working."""
        raise NotImplementedError("MySQL connector not yet implemented")


# Example implementation when you need it:
"""
import pymysql

class MySQLConnector(BaseConnector):
    def connect(self) -> None:
        self.connection = pymysql.connect(
            host=self.config['host'],
            user=self.config['username'],
            password=self.config['password'],
            database=self.config['database'],
            port=self.config.get('port', 3306)
        )
    
    def fetch_last_row(self, table_name: str, order_by: Optional[str] = None) -> Dict[str, Any]:
        cursor = self.connection.cursor(pymysql.cursors.DictCursor)
        if order_by:
            cursor.execute(f"SELECT * FROM {table_name} ORDER BY {order_by} DESC LIMIT 1")
        else:
            cursor.execute(f"SELECT * FROM {table_name} LIMIT 1")
        return cursor.fetchone() or {}
"""
