requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
pyodbc==5.0.1
httpx
google-generativeai
# Optional: Uncomment to enable other LLM providers
# openai
# anthropic